Add support for different scopes and RBAC personas
This WIP idea adds support for using nine specific pre-provisioned user
personas that can be used to test service policies from the point of
view of the three scopes and three default roles that are available out
of the box in keystone. In addition to the os_admin, os_primary, os_alt,
and role-based credentials that were available before, test classes can
now access such credentials as os_system_admin or os_domain_reader.
Currently these users need to be manually created and added to an
accounts.yaml file using the 'type' attribute to distinguish the
personas, but for this to be properly implemented we will need to add
support for generating these users to the account-generator command.

Change-Id: If3823d7d2920486afbcd5de90aa312ef1ad407cc