Add RBAC personas to dynamic creds
The previous commit added nine RBAC personas to the pre-provisioned
credentials provider. This one adds it for the dynamic credentials
provider. This bundles in support for having the [auth] admin be
system-scoped, which should be split into a different commit. We also
need to figure out how to handle the invalid assumptions around domain
scope in a backwards compatible way.

Change-Id: I8acaaece25005163fe26ff7774ab39952538f50d